make it run under darwin
